backend/requirements.txt
tqdm
torch
opencv-python
ffmpegcv
numpy
python_speech_features
scipy
scenedetect
scikit-learn
gdown
torchvision
pandas
transformers
accelerate
datasets
google-genai
pysubs2
boto3
fastapi[standard]
whisperx